# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Swagger",
    "title": "Recipes API Swagger",
    "contact": {},
    "version": "1.0"
  },
  "host": "localhost:8080",
  "paths": {
...
    "/users/me": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get self",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get self",
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/models.UserMe"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to get self",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete self",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Delete self",
        "parameters": [
          {
            "description": "User information to update",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/models.UserMe"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to delete self / delete picture",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update self",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update self",
        "parameters": [
          {
            "description": "User information to update",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/models.UserMe"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "406": {
            "description": "Username too long / invalid email / insecure password",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "409": {
            "description": "User already exists",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to update self",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/users/me/picture": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update your profile picture, must be under 8MB",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update self picture",
        "parameters": [
          {
            "type": "file",
            "description": "Picture, formData must be named 'file'",
            "name": "request",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UpdatePictureResponse"
            }
          },
          "400": {
            "description": "Empty formData / not named correctly",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "406": {
            "description": "Wrong format (not png / jpeg / jpg) / File too heavy",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to: Update user, save the image",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete your profile picture",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Delete own picture",
        "responses": {
          "200": {
            "description": "Picture deleted",
            "schema": {
              "$ref": "#/definitions/models.MessageResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Image not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "422": {
            "description": "No picture for this user",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to: Update user / remove the picture",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "models.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "string",
          "description": "Stable machine-readable error code"
        },
        "request_id": {
          "type": "string",
          "description": "Request identifier, also returned in X-Request-ID"
        }
      }
    },
...
    "models.MessageResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "models.UpdatePictureResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
...
    "models.UpdateUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "picture": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
    "models.UserMe": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "picture": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "basePath": "/api/v1"
}